#54 feat: added depth() to Action trait, moved Index trait to separate module
