+++ src/logic/perft.rs
@@ -5,10 +5,7 @@ use rayon::prelude::*;
 use crate::piece::Piece;
 
 use super::{
-    actions::{play_action, Index},
-    movegen::available_player_actions,
-    rules::{can_move1, can_move2, can_stack, can_unstack, is_action_win},
-    translate::action_to_string,
+    actions::play_action, index::Index, movegen::available_player_actions, rules::{can_move1, can_move2, can_stack, can_unstack, is_action_win}, translate::action_to_string
 };
 
 /// Returns the number of possible actions for a player.
